requirements.txt
# Core dependencies
openai>=1.97.1

numpy==1.26.4
pandas==2.2.0
opencv-python==4.7.0.72
openai>=1.97.1
langchain-perplexity==0.1.2


pyyaml==6.0.1
python-dotenv==1.0.0

# Logging and monitoring
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Utilities
tqdm==4.66.1
tenacity==8.2.3